requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyserial==3.5
pytest==7.4.0
flake8==6.1.0
mypy==1.5.1
paho-mqtt==2.2.1